MorePopover drag UX example
This is a naive and ugly approach to improve the UX of dragging an event
out of a MorePopover in the daygrid.

See examples/daygrid-drag-from-more-popover.html for a demo

This adds `CalendarApi.disableDropOnMorePopover()` which sets the hit
layer of the "More Events" popover to -1 so that it does not intercept
dropping of the event.